客户查询

+++ src/main/kotlin/net/phadata/billing/model/base/PageInfo.kt
@@ -11,18 +11,21 @@ import io.swagger.annotations.ApiModelProperty
  */
 class PageInfo<T : Any> {
     @ApiModelProperty(value = "页码")
-    var current: Long = 0
-    @ApiModelProperty(value = "页数",)
-    var size: Long = 0
-    @ApiModelProperty(value = "总数",)
-    var total: Long = 0
-    @ApiModelProperty(value = "数据",)
+    var current: Int = 0
+
+    @ApiModelProperty(value = "页数")
+    var size: Int = 0
+
+    @ApiModelProperty(value = "总数")
+    var total: Int = 0
+
+    @ApiModelProperty(value = "数据")
     var records: List<T>? = null
     fun buildPageInfo(page: Page<T>): PageInfo<T> {
         return PageInfo<T>().apply {
-            current = page.current
-            size = page.size
-            total = page.total
+            current = page.current.toInt()
+            size = page.size.toInt()
+            total = page.total.toInt()
             records = page.records
         }
     }
